Shipit: add Importit PoC
It doesn't do much yet - it just prints the imported PR into console. I think it's gonna be enough to just import it into local branch so our devs can work with it. I also changed the source/destination terminology because it's quite confusing (source and destination change depending on the direction).

kiwicom-source-id: ce265ba012a1174aa6c30a4cddbe17700d52dc88