Monorepo Shipit: fail lazily during projects iteration
Previously we failed the process immediately when one export failed. This however blocks other exports which are fine. Now, we are going to collect these errors and fail only after we iterate all the projects. This way we can continue successfully shipping these projects while having the job fail if some of them were unsuccessful.

kiwicom-source-id: 854b92f65a7162f2c0521f4a66a842503d685b7d