Remove platform repositories
Platform team doesn't want to use this anymore from what I understood (based on our conversation in #plz-incubator and this repo which is already outside of Shipit: https://github.com/kiwicom/terraform-google-stackdriver-logs-exclusion). Every repository is available on GitHub which is their primary destination anyway. Therefore no code is lost. As always: we can easily revert it.

kiwicom-source-id: 418c1401e43a97a3f05d3fe68b8373cdd08b62e4